% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_line_plot.R
\name{bd_line_plot}
\alias{bd_line_plot}
\title{Visualize spatial clustering and expression data distribution associated with the boundary.}
\usage{
bd_line_plot(
  st = stobj,
  group.by = NULL,
  features = NULL,
  slot = "data",
  min.cutoff = NA,
  max.cutoff = NA,
  dis.group = c(-Inf, -3, -2, -1, 0, 1, 2, 3, Inf),
  sample_id = "orig.ident",
  subgroup = NULL,
  subgroup_type = NULL,
  ylim = NULL,
  method = "median",
  col = NULL,
  shape = 21,
  pt.size = 0,
  title = "Spatial distribution",
  title.size = 12
)
}
\arguments{
\item{st}{Spatial transcriptomics seurat object.}

\item{group.by}{The categorical or numeric features for the spatial distribution visualization.}

\item{features}{Multiple numeric features for the spatial distribution visualization.}

\item{slot}{If plotting a feature, which data slot to pull from (counts, data, or scale.data).}

\item{min.cutoff}{Vector of minimum and maximum cutoff values for each feature, may specify quantile in the form of 'q##' where '##' is the quantile (eg, 'q1', 'q10')}

\item{max.cutoff}{See min.cutoff}

\item{dis.group}{The distance group. A vector to split the spatial distance associated with boundary (e.g. c(-Inf,-4,-3,-2,-1,0,1,2,3,4,Inf)).}

\item{sample_id}{If multiple sample is merged in the st object, the column of different sample_id should be filled in here. By default, sample_id = 'orig.ident'.}

\item{subgroup}{If only a subset of the st object is used to plot, fill in the column name of the categorical variable here and the spots keep in for "subgroup_type".}

\item{subgroup_type}{See subgroup (NULL for default).}

\item{ylim}{Y axis limits of the plot (NULL for default).}

\item{method}{Use the mean or median of the features value among spots in certain distance group. By default, method = "median". "mean" can also be used when the features conforms to a normal distribution.}

\item{col}{colors used to plot (NULL for default).}

\item{shape}{Control the shape of the spots - same as the ggplot2 parameter. The default is 21, which plots circles - use 22 to plot squares.}

\item{pt.size}{Size of the spots. The default is 0.}

\item{title}{Title of the plot.}

\item{title.size}{Text size of the title.}
}
\value{
Line plot of features or discrete grouping (e.g. cluster assignments) distribution associated with the boundary.
}
\description{
Visualize spatial clustering and expression data distribution associated with the boundary.
}
\examples{
plot <- bd_line_plot(st = stobj,
                               group.by = NULL,
                               features = NULL,
                               slot = 'data',
                               min.cutoff = NA,
                               max.cutoff = NA,
                               dis.group = c(-Inf,-3,-2,-1,0,1,2,3,Inf),
                               sample_id = 'orig.ident',
                               subgroup = NULL,
                               subgroup_type = NULL,
                               ylim = NULL,
                               method = 'median',
                               col = NULL,
                               shape = 21,
                               pt.size = 0,
                               title = 'Spatial distribution',
                               title.size = 12)
}
