% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_posst_score.R
\name{calculate_posst_score}
\alias{calculate_posst_score}
\title{Create the Matrix for Calculating Neighborhood/Aggregation/Co-localization Score}
\usage{
calculate_posst_score(
  st = stobj,
  niche = "niche",
  images = names(st@images),
  platform = "10x"
)
}
\arguments{
\item{st}{Spatial transcriptomics seurat object.}

\item{niche}{Spatial niches or clusters.}

\item{images}{Vector of the name for all the images. By default, all the slides are included.}

\item{platform}{Platform of spatial transcriptomics. 10x visium or BMK s1000 (L13 bin) spatial transcriptomics can be loaded as input for this package. Corresponding values are "10x" and "s1000". By default, platform = '10x'.}
}
\value{
The total number of spots around the center spot and neighborhood/aggregation/co-localization score. Corresponding values are "total_neighbor_count", "neigh_score", "agg_score", "coloc_score_..."
}
\description{
Create the Matrix for Calculating Neighborhood/Aggregation/Co-localization Score
}
\examples{
images <- names(stobj@images)
st <- calculate_posst_score(st = stobj,
                            niche = 'niche',
                            images = images,
                            platform = 10x)
}
