% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_posst_boundary.R
\name{find_posst_boundary}
\alias{find_posst_boundary}
\title{Find the boundary between the center niche and other niches}
\usage{
find_posst_boundary(
  st = stobj,
  niche = "niche",
  center_niche = "niche_1",
  images = names(st@images),
  strictness = 2,
  platform = "10x"
)
}
\arguments{
\item{st}{Spatial transcriptomics seurat object.}

\item{niche}{Spatial niches or clusters.}

\item{center_niche}{The center niche.}

\item{images}{Vector of the name for all the images. By default, all the slides are included.}

\item{strictness}{Number of neighboring spots in boundary finding, indicating the strictness of the boundary definition. This parameter can take the values 1, 2, or 3, with 3 representing the strictest setting. (By default strictness = 2)}

\item{platform}{Platform of spatial transcriptomics. 10x visium or BMK s1000 (L13 bin) spatial transcriptomics can be loaded as input for this package. Corresponding values are "10x" and "s1000". By default, platform = '10x'.}
}
\value{
The boundary spots between the center niche and other niches were labeled in "boundary". The minimal spatial distance for each spots to the boundary ("MDB" for short) were calculated. "MDBu" represents the value adjusted from MDB, which holds practical significance and is measured in units of 100 micrometers (μm).
}
\description{
Find the boundary between the center niche and other niches
}
\examples{
st <- find_posst_boundary(st = stobj,
                                    niche = 'niche',
                                    center_niche = 'niche_1',
                                    images = names(st@images),
                                    strictness = 2,
                                    platform = '10x')
}
