% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posst_coloc_heatmap.R
\name{posst_coloc_heatmap}
\alias{posst_coloc_heatmap}
\title{Co-localization score heatmap of all the niches}
\usage{
posst_coloc_heatmap(
  st = stobj,
  niche = "niche",
  subgroup = NULL,
  subgroup_type = NULL,
  col = rev(brewer.pal(n = 7, name = "RdYlGn")),
  scale = "none",
  title = "niche co-localization score"
)
}
\arguments{
\item{st}{Spatial transcriptomics seurat object (after Run create_posst_matrix function).}

\item{niche}{Spatial niches or clusters.}

\item{subgroup}{a factor in object metadata to subset the stobj (e.g. the treatment options, or the sampling sites. NULL for default).}

\item{subgroup_type}{the a value of subgroup to choose (e.g. treat/control, or tumor/para-tumor. Single value or a vector of values can be set as the input. NULL for default).}

\item{col}{colors used to plot (rev(brewer.pal(n = 7, name ="RdYlGn") for default).}

\item{scale}{character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. Corresponding values are "row","column", and "none".}

\item{title}{Title of the plot.}
}
\value{
The Co-localization heatmap of all the niches, calculating with the POSST package.
}
\description{
Co-localization score heatmap of all the niches
}
\examples{
plot <- posst_coloc_heatmap(st = stobj,
                                      niche = 'niche',
                                      subgroup = 'sampling_site',
                                      subgroup_type = c('T','PT'),
                                      col = rev(brewer.pal(n = 7, name ="RdYlGn")),
                                      scale = 'none',
                                      title = 'Niche co-localization score')
}
