% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posst_coloc_lollipop_plot.R
\name{posst_coloc_lollipop_plot}
\alias{posst_coloc_lollipop_plot}
\title{Co-localization score lollipop plot between the center niche and other niches}
\usage{
posst_coloc_lollipop_plot(
  st = stobj,
  niche = "niche",
  center_niche = "niche_1",
  subgroup = NULL,
  subgroup_type = NULL,
  col = NULL,
  dot.size = 5,
  ylim = c(0, 6),
  label.size = 3,
  title = paste("Co-localization score between", center_niche, "and other niches",
    subgroup_type, sep = " "),
  title.size = 12
)
}
\arguments{
\item{st}{Spatial transcriptomics seurat object (after Run create_posst_matrix function).}

\item{niche}{Spatial niches or clusters.}

\item{center_niche}{The center niche.}

\item{subgroup}{a factor in object metadata to subset the stobj (e.g. the treatment options, or the sampling sites. NULL for default).}

\item{subgroup_type}{the a value of subgroup to choose (e.g. treat/control, or tumor/para-tumor. Single value or a vector of values can be set as the input. NULL for default).}

\item{col}{colors used to plot (NULL for default).}

\item{dot.size}{dot size of the lollipop. By default the dot.size is 5.}

\item{ylim}{Y axis limits of the plot.}

\item{label.size}{Text size of labels.}

\item{title}{Title of the plot.}

\item{title.size}{Text size of the title.}
}
\value{
The lollipop plot of co-localization score between the center niche X and other niches calculated with the POSST package.
}
\description{
Co-localization score lollipop plot between the center niche and other niches
}
\examples{
plot <- posst_coloc_lollipop_plot(st = stobj,
                                            niche = 'niche',
                                            center_niche = 'niche_1',
                                            subgroup = 'sampling_site',
                                            subgroup = c('T','PT'),
                                            col = c('niche1' = 'f60000',...),
                                            dot.size = 5,
                                            ylim = c(0,6),
                                            label.size = 3,
                                            title = paste('Co-localization score between',center_niche,'and other niches',subgroup_type,sep = ' '),
                                            title.size = 12)
}
